requirements.txt
flask
flask_httpauth
rembg
pillow
onnxruntime
waitress
python-dotenv